src/requirements.txt
waitress==1.4.3
flask==1.1.2
flask-restplus==0.13.0
flask-jwt-extended==3.24.1
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
python-gammu==2.12
mysqlclient==1.4.6

# Fix bug in flask
werkzeug==0.16.1
